split huge script into multiple files
