+++ colors.py
@@ -1,0 +1,12 @@
+class Colors:
+    HEADER = '\033[95m'
+    BLUE = '\033[94m'
+    CYAN = '\033[96m'
+    GREEN = '\033[92m'
+    PURPLE = '\033[95m'   # For loading/progress/waiting messages
+    ORANGE = '\033[38;5;208m'
+    WARNING = '\033[95m'  # Use purple for warnings
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
